+++ GHEtool/gui/test_gui/test_results_DHW.py
@@ -15,6 +15,8 @@ def test_results_DHW(qtbot):
                              data_2_results_function=data_2_borefield)
 
     gs = main_window.gui_structure
+    gs.option_source_ground_temperature.set_value(0)
+    gs.option_method_rb_calc.set_value(0)
 
     gs.aim_optimize.widget.click()
     main_window.gui_structure.option_decimal_csv.set_value(0)
@@ -33,7 +35,7 @@ def test_results_DHW(qtbot):
 
     main_window.display_results()
 
-    assert gs.results_heating_load.label.text() == "Heating load on the borefield: 566750.0 kWh"
+    assert gs.results_heating_load.label.text() == "Heating load on the borefield: 530633.0 kWh"
 
     gs.aim_temp_profile.widget.click()
     gs.geo_load.set_value(0)
@@ -129,5 +131,5 @@ def test_results_DHW(qtbot):
 
     main_window.display_results()
 
-    assert gs.results_heating_load.label.text() == "Heating load on the borefield: 566750.0 kWh"
+    assert gs.results_heating_load.label.text() == "Heating load on the borefield: 530633.0 kWh"